witchy-booktracker-starter/backend/requirements.txt

Flask
Flask-JWT-Extended
Flask-CORS
SQLAlchemy
psycopg2-binary
python-dotenv
requests
werkzeug
